splits indexing methods into new module
